// src/services/obsidianRestAPI/generated-types.ts
export interface components {
  schemas: {
    Error: {
      /**
       * @description A 5-digit error code uniquely identifying this particular type of error.
       *
       * @example 40149
       */
      errorCode?: number;
      /**
       * @description Message describing the error.
       * @example A brief description of the error.
       */
      message?: string;
    };
    NoteJson: {
      content: string;
      frontmatter: {
        [key: string]: unknown;
      };
      path: string;
      stat: {
        ctime: number;
        mtime: number;
        size: number;
      };
      tags: string[];
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}